# de-logi-hub-web/src/app/core/api/models/register-user-model.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Role } from '../models/role';
export interface RegisterUserModel {
  address: string;
  birthDate: string;
  email: string;
  firstName: string;
  lastName: string;
  login: string;
  password: string;
  phoneNumber: string;
  profilePicture?: string | null;
  role: Role;
  userId: string;
}