# de-logi-hub-web/src/app/core/api/models/transport-type.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export enum TransportType {
  Unknown = 'Unknown',
  None = 'None',
  Truck = 'Truck',
  Van = 'Van',
  Refrigerated = 'Refrigerated',
  Tanker = 'Tanker',
  Container = 'Container'
}