# de-logi-hub-web/src/app/core/api/models/user-profile-model.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Role } from '../models/role';
export interface UserProfileModel {
  address?: string | null;
  birthDate?: string;
  birthDateString?: string | null;
  email?: string | null;
  firstName?: string | null;
  id?: string;
  lastName?: string | null;
  login?: string | null;
  phoneNumber?: string | null;
  profilePicture?: string | null;
  registrationDate?: string;
  registrationDateStr?: string | null;
  role?: Role;
}