# de-logi-hub-web/src/openapi/openapi.json
{
	"openapi": "3.0.1",
	"info": {
	  "title": "DeLogiHubApi",
	  "version": "1.0"
	},
	"paths": {
	  "/api/auth/login": {
		"post": {
		  "tags": [
			"Auth"
		  ],
		  "operationId": "Login",
		  "requestBody": {
			"content": {
			  "application/json": {
				"schema": {
				  "$ref": "#/components/schemas/LoginModel"
				}
			  }
			}
		  },
		  "responses": {
			"200": {
			  "description": "Success",
			  "content": {
				"application/json": {
				  "schema": {
					"$ref": "#/components/schemas/Token"
				  }
				}
			  }
			}
		  }
		}
	  },
	  "/api/auth/refresh": {
		"get": {
		  "tags": [
			"Auth"
		  ],
		  "operationId": "Refresh",
		  "parameters": [
			{
			  "name": "refreshTokenString",
			  "in": "header",
			  "schema": {
				"type": "string"
			  }
			}
		  ],
		  "responses": {
			"200": {
			  "description": "Success",
			  "content": {
				"application/json": {
				  "schema": {
					"$ref": "#/components/schemas/Token"
				  }
				}
			  }
			}
		  }
		}
	  },
	  "/api/offers/add": {
		"post": {
		  "tags": [
			"Offers"
		  ],
		  "operationId": "AddOffer",
		  "requestBody": {
			"content": {
			  "application/json": {
				"schema": {
				  "$ref": "#/components/schemas/OfferModel"
				}
			  }
			}
		  },
		  "responses": {
			"200": {
			  "description": "Success"
			},
			"401": {
			  "description": "Unauthorized"
			},
			"403": {
			  "description": "Forbidden"
			}
		  },
		  "security": [
			{
			  "oauth2": [ ]
			}
		  ]
		}
	  },
	  "/api/offers/update": {
		"post": {
		  "tags": [
			"Offers"
		  ],
		  "operationId": "UpdateOffer",
		  "requestBody": {
			"content": {
			  "application/json": {
				"schema": {
				  "$ref": "#/components/schemas/OfferModel"
				}
			  }
			}
		  },
		  "responses": {
			"200": {
			  "description": "Success"
			},
			"401": {
			  "description": "Unauthorized"
			},
			"403": {
			  "description": "Forbidden"
			}
		  },
		  "security": [
			{
			  "oauth2": [ ]
			}
		  ]
		}
	  },
	  "/api/offers/delete": {
		"delete": {
		  "tags": [
			"Offers"
		  ],
		  "operationId": "DeleteOffer",
		  "parameters": [
			{
			  "name": "id",
			  "in": "query",
			  "schema": {
				"type": "string",
				"format": "uuid"
			  }
			}
		  ],
		  "responses": {
			"200": {
			  "description": "Success"
			},
			"401": {
			  "description": "Unauthorized"
			},
			"403": {
			  "description": "Forbidden"
			}
		  },
		  "security": [
			{
			  "oauth2": [ ]
			}
		  ]
		}
	  },
	  "/api/offers/get": {
		"get": {
		  "tags": [
			"Offers"
		  ],
		  "operationId": "GetOffer",
		  "parameters": [
			{
			  "name": "id",
			  "in": "query",
			  "schema": {
				"type": "string",
				"format": "uuid"
			  }
			}
		  ],
		  "responses": {
			"200": {
			  "description": "Success",
			  "content": {
				"application/json": {
				  "schema": {
					"$ref": "#/components/schemas/OfferModel"
				  }
				}
			  }
			},
			"401": {
			  "description": "Unauthorized"
			},
			"403": {
			  "description": "Forbidden"
			}
		  },
		  "security": [
			{
			  "oauth2": [ ]
			}
		  ]
		}
	  },
	  "/api/offers/get-all-by-order": {
		"get": {
		  "tags": [
			"Offers"
		  ],
		  "operationId": "GetOffersByOrder",
		  "parameters": [
			{
			  "name": "orderId",
			  "in": "query",
			  "schema": {
				"type": "string",
				"format": "uuid"
			  }
			}
		  ],
		  "responses": {
			"200": {
			  "description": "Success",
			  "content": {
				"application/json": {
				  "schema": {
					"type": "array",
					"items": {
					  "$ref": "#/components/schemas/OfferModel"
					}
				  }
				}
			  }
			},
			"401": {
			  "description": "Unauthorized"
			},
			"403": {
			  "description": "Forbidden"
			}
		  },
		  "security": [
			{
			  "oauth2": [ ]
			}
		  ]
		}
	  },
	  "/api/offers/get-all-by-carrier": {
		"get": {
		  "tags": [
			"Offers"
		  ],
		  "operationId": "GetOffersByCarrier",
		  "parameters": [
			{
			  "name": "carrierId",
			  "in": "query",
			  "schema": {
				"type": "string",
				"format": "uuid"
			  }
			}
		  ],
		  "responses": {
			"200": {
			  "description": "Success",
			  "content": {
				"application/json": {
				  "schema": {
					"type": "array",
					"items": {
					  "$ref": "#/components/schemas/OfferModel"
					}
				  }
				}
			  }
			},
			"401": {
			  "description": "Unauthorized"
			},
			"403": {
			  "description": "Forbidden"
			}
		  },
		  "security": [
			{
			  "oauth2": [ ]
			}
		  ]
		}
	  },
	  "/api/orders/get-all-by-customer": {
		"get": {
		  "tags": [
			"Orders"
		  ],
		  "operationId": "GetCustomerOrders",
		  "parameters": [
			{
			  "name": "customerId",
			  "in": "query",
			  "schema": {
				"type": "string",
				"format": "uuid"
			  }
			}
		  ],
		  "responses": {
			"200": {
			  "description": "Success",
			  "content": {
				"application/json": {
				  "schema": {
					"type": "array",
					"items": {
					  "$ref": "#/components/schemas/OrderModel"
					}
				  }
				}
			  }
			},
			"401": {
			  "description": "Unauthorized"
			},
			"403": {
			  "description": "Forbidden"
			}
		  },
		  "security": [
			{
			  "oauth2": [ ]
			}
		  ]
		}
	  },
	  "/api/orders/get-all": {
		"get": {
		  "tags": [
			"Orders"
		  ],
		  "operationId": "GetOrders",
		  "responses": {
			"200": {
			  "description": "Success",
			  "content": {
				"application/json": {
				  "schema": {
					"type": "array",
					"items": {
					  "$ref": "#/components/schemas/OrderModel"
					}
				  }
				}
			  }
			},
			"401": {
			  "description": "Unauthorized"
			},
			"403": {
			  "description": "Forbidden"
			}
		  },
		  "security": [
			{
			  "oauth2": [ ]
			}
		  ]
		}
	  },
	  "/api/orders/get": {
		"get": {
		  "tags": [
			"Orders"
		  ],
		  "operationId": "GetOrder",
		  "parameters": [
			{
			  "name": "id",
			  "in": "query",
			  "schema": {
				"type": "string",
				"format": "uuid"
			  }
			}
		  ],
		  "responses": {
			"200": {
			  "description": "Success",
			  "content": {
				"application/json": {
				  "schema": {
					"$ref": "#/components/schemas/OrderModel"
				  }
				}
			  }
			},
			"401": {
			  "description": "Unauthorized"
			},
			"403": {
			  "description": "Forbidden"
			}
		  },
		  "security": [
			{
			  "oauth2": [ ]
			}
		  ]
		}
	  },
	  "/api/orders/add": {
		"post": {
		  "tags": [
			"Orders"
		  ],
		  "operationId": "AddOrder",
		  "requestBody": {
			"content": {
			  "application/json": {
				"schema": {
				  "$ref": "#/components/schemas/OrderModel"
				}
			  }
			}
		  },
		  "responses": {
			"200": {
			  "description": "Success"
			},
			"401": {
			  "description": "Unauthorized"
			},
			"403": {
			  "description": "Forbidden"
			}
		  },
		  "security": [
			{
			  "oauth2": [ ]
			}
		  ]
		}
	  },
	  "/api/orders/update": {
		"post": {
		  "tags": [
			"Orders"
		  ],
		  "operationId": "UpdateOrder",
		  "requestBody": {
			"content": {
			  "application/json": {
				"schema": {
				  "$ref": "#/components/schemas/OrderModel"
				}
			  }
			}
		  },
		  "responses": {
			"200": {
			  "description": "Success"
			},
			"401": {
			  "description": "Unauthorized"
			},
			"403": {
			  "description": "Forbidden"
			}
		  },
		  "security": [
			{
			  "oauth2": [ ]
			}
		  ]
		}
	  },
	  "/api/orders/delete": {
		"delete": {
		  "tags": [
			"Orders"
		  ],
		  "operationId": "DeleteOrder",
		  "parameters": [
			{
			  "name": "id",
			  "in": "query",
			  "schema": {
				"type": "string",
				"format": "uuid"
			  }
			}
		  ],
		  "responses": {
			"200": {
			  "description": "Success"
			},
			"401": {
			  "description": "Unauthorized"
			},
			"403": {
			  "description": "Forbidden"
			}
		  },
		  "security": [
			{
			  "oauth2": [ ]
			}
		  ]
		}
	  },
	  "/api/transports/get-all-by-carrier": {
		"get": {
		  "tags": [
			"Transports"
		  ],
		  "operationId": "GetTransportsByCarrier",
		  "parameters": [
			{
			  "name": "carrierId",
			  "in": "query",
			  "schema": {
				"type": "string",
				"format": "uuid"
			  }
			}
		  ],
		  "responses": {
			"200": {
			  "description": "Success",
			  "content": {
				"application/json": {
				  "schema": {
					"type": "array",
					"items": {
					  "$ref": "#/components/schemas/TransportModel"
					}
				  }
				}
			  }
			},
			"401": {
			  "description": "Unauthorized"
			},
			"403": {
			  "description": "Forbidden"
			}
		  },
		  "security": [
			{
			  "oauth2": [ ]
			}
		  ]
		}
	  },
	  "/api/transports/get": {
		"get": {
		  "tags": [
			"Transports"
		  ],
		  "operationId": "GetTransport",
		  "parameters": [
			{
			  "name": "id",
			  "in": "query",
			  "schema": {
				"type": "string",
				"format": "uuid"
			  }
			}
		  ],
		  "responses": {
			"200": {
			  "description": "Success",
			  "content": {
				"application/json": {
				  "schema": {
					"$ref": "#/components/schemas/TransportModel"
				  }
				}
			  }
			},
			"401": {
			  "description": "Unauthorized"
			},
			"403": {
			  "description": "Forbidden"
			}
		  },
		  "security": [
			{
			  "oauth2": [ ]
			}
		  ]
		}
	  },
	  "/api/transports/add": {
		"post": {
		  "tags": [
			"Transports"
		  ],
		  "operationId": "AddTransport",
		  "requestBody": {
			"content": {
			  "application/json": {
				"schema": {
				  "$ref": "#/components/schemas/TransportModel"
				}
			  }
			}
		  },
		  "responses": {
			"200": {
			  "description": "Success"
			},
			"401": {
			  "description": "Unauthorized"
			},
			"403": {
			  "description": "Forbidden"
			}
		  },
		  "security": [
			{
			  "oauth2": [ ]
			}
		  ]
		}
	  },
	  "/api/transports/update": {
		"post": {
		  "tags": [
			"Transports"
		  ],
		  "operationId": "UpdateTransport",
		  "requestBody": {
			"content": {
			  "application/json": {
				"schema": {
				  "$ref": "#/components/schemas/TransportModel"
				}
			  }
			}
		  },
		  "responses": {
			"200": {
			  "description": "Success"
			},
			"401": {
			  "description": "Unauthorized"
			},
			"403": {
			  "description": "Forbidden"
			}
		  },
		  "security": [
			{
			  "oauth2": [ ]
			}
		  ]
		}
	  },
	  "/api/transports/delete": {
		"delete": {
		  "tags": [
			"Transports"
		  ],
		  "operationId": "DeleteTransport",
		  "parameters": [
			{
			  "name": "id",
			  "in": "query",
			  "schema": {
				"type": "string",
				"format": "uuid"
			  }
			}
		  ],
		  "responses": {
			"200": {
			  "description": "Success"
			},
			"401": {
			  "description": "Unauthorized"
			},
			"403": {
			  "description": "Forbidden"
			}
		  },
		  "security": [
			{
			  "oauth2": [ ]
			}
		  ]
		}
	  },
	  "/api/users/register": {
		"post": {
		  "tags": [
			"Users"
		  ],
		  "operationId": "Register",
		  "requestBody": {
			"content": {
			  "application/json": {
				"schema": {
				  "$ref": "#/components/schemas/RegisterUserModel"
				}
			  }
			}
		  },
		  "responses": {
			"200": {
			  "description": "Success"
			}
		  }
		}
	  },
	  "/api/users/update": {
		"post": {
		  "tags": [
			"Users"
		  ],
		  "operationId": "UpdateUserProfile",
		  "requestBody": {
			"content": {
			  "application/json": {
				"schema": {
				  "$ref": "#/components/schemas/UserProfileInfo"
				}
			  }
			}
		  },
		  "responses": {
			"200": {
			  "description": "Success"
			},
			"401": {
			  "description": "Unauthorized"
			},
			"403": {
			  "description": "Forbidden"
			}
		  },
		  "security": [
			{
			  "oauth2": [ ]
			}
		  ]
		}
	  },
	  "/api/users/get": {
		"get": {
		  "tags": [
			"Users"
		  ],
		  "operationId": "GetUserProfileById",
		  "parameters": [
			{
			  "name": "id",
			  "in": "query",
			  "schema": {
				"type": "string",
				"format": "uuid"
			  }
			}
		  ],
		  "responses": {
			"200": {
			  "description": "Success",
			  "content": {
				"application/json": {
				  "schema": {
					"$ref": "#/components/schemas/UserProfileModel"
				  }
				}
			  }
			},
			"401": {
			  "description": "Unauthorized"
			},
			"403": {
			  "description": "Forbidden"
			}
		  },
		  "security": [
			{
			  "oauth2": [ ]
			}
		  ]
		}
	  }
	},
	"components": {
	  "schemas": {
		"LoginModel": {
		  "type": "object",
		  "properties": {
			"login": {
			  "type": "string",
			  "nullable": true
			},
			"password": {
			  "type": "string",
			  "nullable": true
			}
		  },
		  "additionalProperties": false
		},
		"OfferModel": {
		  "type": "object",
		  "properties": {
			"id": {
			  "type": "string",
			  "format": "uuid"
			},
			"carrierId": {
			  "type": "string",
			  "format": "uuid"
			},
			"orderId": {
			  "type": "string",
			  "format": "uuid"
			},
			"status": {
			  "$ref": "#/components/schemas/OfferStatus"
			}
		  },
		  "additionalProperties": false
		},
		"OfferStatus": {
		  "enum": [
			"Unknown",
			"Pending",
			"Accepted",
			"Rejected"
		  ],
		  "type": "string"
		},
		"OrderModel": {
		  "type": "object",
		  "properties": {
			"id": {
			  "type": "string",
			  "format": "uuid"
			},
			"customerId": {
			  "type": "string",
			  "format": "uuid"
			}
		  },
		  "additionalProperties": false
		},
		"RegisterUserModel": {
		  "required": [
			"address",
			"birthDate",
			"email",
			"firstName",
			"lastName",
			"login",
			"password",
			"phoneNumber",
			"role",
			"userId"
		  ],
		  "type": "object",
		  "properties": {
			"userId": {
			  "type": "string",
			  "format": "uuid"
			},
			"login": {
			  "minLength": 1,
			  "type": "string"
			},
			"password": {
			  "minLength": 1,
			  "type": "string"
			},
			"firstName": {
			  "minLength": 1,
			  "type": "string"
			},
			"lastName": {
			  "minLength": 1,
			  "type": "string"
			},
			"profilePicture": {
			  "type": "string",
			  "format": "byte",
			  "nullable": true
			},
			"address": {
			  "minLength": 1,
			  "type": "string"
			},
			"phoneNumber": {
			  "minLength": 1,
			  "type": "string"
			},
			"birthDate": {
			  "type": "string",
			  "format": "date-time"
			},
			"email": {
			  "minLength": 1,
			  "type": "string"
			},
			"role": {
			  "$ref": "#/components/schemas/Role"
			}
		  },
		  "additionalProperties": false
		},
		"Role": {
		  "enum": [
			"Customer",
			"Carrier"
		  ],
		  "type": "string"
		},
		"Token": {
		  "type": "object",
		  "properties": {
			"accessToken": {
			  "type": "string",
			  "nullable": true
			},
			"refreshToken": {
			  "type": "string",
			  "nullable": true
			}
		  },
		  "additionalProperties": false
		},
		"TransportModel": {
		  "type": "object",
		  "properties": {
			"id": {
			  "type": "string",
			  "format": "uuid"
			},
			"type": {
			  "$ref": "#/components/schemas/TransportType"
			},
			"licensePlate": {
			  "type": "string",
			  "nullable": true
			},
			"capacity": {
			  "type": "number",
			  "format": "double"
			},
			"carrierId": {
			  "type": "string",
			  "format": "uuid"
			}
		  },
		  "additionalProperties": false
		},
		"TransportType": {
		  "enum": [
			"Unknown",
			"None",
			"Truck",
			"Van",
			"Refrigerated",
			"Tanker",
			"Container"
		  ],
		  "type": "string"
		},
		"UserProfileInfo": {
		  "type": "object",
		  "properties": {
			"id": {
			  "type": "string",
			  "format": "uuid"
			},
			"firstName": {
			  "type": "string",
			  "nullable": true
			},
			"lastName": {
			  "type": "string",
			  "nullable": true
			},
			"profilePicture": {
			  "type": "string",
			  "format": "byte",
			  "nullable": true
			},
			"address": {
			  "type": "string",
			  "nullable": true
			},
			"phoneNumber": {
			  "type": "string",
			  "nullable": true
			},
			"birthDate": {
			  "type": "string",
			  "format": "date-time"
			},
			"email": {
			  "type": "string",
			  "nullable": true
			}
		  },
		  "additionalProperties": false
		},
		"UserProfileModel": {
		  "type": "object",
		  "properties": {
			"id": {
			  "type": "string",
			  "format": "uuid"
			},
			"firstName": {
			  "type": "string",
			  "nullable": true
			},
			"lastName": {
			  "type": "string",
			  "nullable": true
			},
			"profilePicture": {
			  "type": "string",
			  "format": "byte",
			  "nullable": true
			},
			"address": {
			  "type": "string",
			  "nullable": true
			},
			"phoneNumber": {
			  "type": "string",
			  "nullable": true
			},
			"birthDate": {
			  "type": "string",
			  "format": "date-time"
			},
			"birthDateString": {
			  "type": "string",
			  "nullable": true
			},
			"email": {
			  "type": "string",
			  "nullable": true
			},
			"login": {
			  "type": "string",
			  "nullable": true
			},
			"registrationDate": {
			  "type": "string",
			  "format": "date-time"
			},
			"registrationDateStr": {
			  "type": "string",
			  "nullable": true
			},
			"role": {
			  "$ref": "#/components/schemas/Role"
			}
		  },
		  "additionalProperties": false
		}
	  },
	  "securitySchemes": {
		"oauth2": {
		  "type": "apiKey",
		  "description": "Standard Authorization header using the Bearer scheme (\"bearer {token}\")",
		  "name": "Authorization",
		  "in": "header"
		}
	  }
	}
  }